fix(ci): python-coverage-shards — split oversized test_adapters.py to satisfy maintainability line-limit guardrail
test_adapters.py reached 1514 lines, tripping the
test_first_party_code_files_stay_under_line_limit guardrail (max 1500).
Move TestOpenCodeAdapter into a new tests/unit/adapters/test_opencode_adapter.py
that imports shared helpers from .test_adapters, matching the existing
per-adapter file pattern (cursor, grok). No test behavior changed.
